Update zn.sh
When I tried running the script as it was, I ran into several issues. 
1. gcc had been replaced by clang
2. the pip2 install for gevent/greenlet/msgpack wasn't working due to an error with greenlet. 
3. The directory that was originally being drawn in wouldn't actually let me run Zeronet as I kept getting an error when the python2 zeronet.py was started. 

I found work arounds for all the issues and then pulled in the master.tar.gz suggested for the zeronet Debian install instructions. Otherwise all the rest of the script was kept the same. There are still some issues with specific sites when running Zeronet through this Termux setup that I haven't had when setting it up using GNURoot. However, the installation will run and you can access Zeronet now.